+++ src/pages/auth/login.tsx
@@ -1,0 +1,64 @@
+import React from 'react';
+import type { NextPage } from 'next';
+import axios from 'axios';
+import Head from 'next/head';
+import { useRouter } from 'next/router';
+import { useSnackbar } from 'notistack';
+import { useForm } from '@hooks';
+import { ICredentials } from '@interfaces';
+
+const initialValues: ICredentials = {
+  email: '',
+  password: '',
+};
+
+const Login: NextPage = () => {
+  const [formValues, handleChange] = useForm<ICredentials>(initialValues);
+  const { email, password } = formValues;
+  const router = useRouter();
+  const { enqueueSnackbar } = useSnackbar();
+
+  const handleSubmit = async (e: React.MouseEvent<HTMLButtonElement>) => {
+    e.preventDefault();
+    try {
+      const response = await axios.post('/api/auth/login', { email, password });
+      if (response.status === 200) {
+        router.push('/admin/dashboard');
+      }
+    } catch (error) {
+      enqueueSnackbar(`${error}`, { variant: 'error' });
+    }
+  };
+
+  return (
+    <div>
+      <Head>
+        <title>My Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <h1>Login</h1>
+      <form>
+        <input
+          type="email"
+          name="email"
+          onChange={handleChange}
+          placeholder="email"
+          value={email}
+        />
+        <input
+          type="password"
+          name="password"
+          onChange={handleChange}
+          placeholder="password"
+          value={password}
+        />
+        <button type="submit" onClick={handleSubmit}>
+          Login
+        </button>
+      </form>
+    </div>
+  );
+};
+
+export default Login;
